refactor: move the custom property flatten helper to a domain util file
flattenOrganizationRepositoryCustomPropertyDefaultValue is used by both the
resource and the data source, so per review feedback it moves out of the
resource file into github/util_custom_property.go, following the
util_<domain>.go convention in ARCHITECTURE.md.

buildOrganizationRepositoryCustomProperty and the two value-type var blocks
stay put — they are only used by the resource.
